Initial commit. Headers for vision and robot communication added.
vision.h contains the headers for the vision processing system. This includes a
see() method, which returns a BallInfo structure. Accessor methods are provided
to extract data on the spatial position of the ball.

robot-comms.h contains the headers for the robot communication system.
setMotorSpeeds allows the speeds of the individual motors to be set, while
setRobotCourse controls the motors to make the robot follow a circular path.